test(executor): make executor setup generic
This makes the setup function reusable for most (all?) executor tests.
